requirements.txt
<<<<<<< HEAD
# apt install liblzma-dev pkg-config
#
# quando o polyglot estiver implementado:
# apt install libicu-dev python3-icu
# after installation:
#   polyglot download embeddings2.en sentiment2.en ner2.en
#   polyglot download embeddings2.pt sentiment2.pt ner2.pt

Django<4
mysqlclient==2.1.0
git+https://bitbucket.org/josir/poweradmin.git@django32
django-crispy-forms
django-select2
django-smart-selects
pytz
tweepy
searchtweets-v2
openpyxl
django-admin-tools
numpy
wordcloud
matplotlib
networkx
seaborn
sqlparse
reportlab==3.6.6
plotly
gunicorn
boto3
